tsc: have far more accurracy

+++ base/usr/include/kernel/arch/x86_64/pit.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <stdint.h>
 
-void pit_oneshot(uint16_t microseconds);
+void pit_oneshot(uint64_t microseconds);
